WFLY-3723: Cleanup, extract Locale, low-level parsing to separate utility class
